refactor: Extract component from InstallationPage
Will be easier for tests
